requirements.txt
Django==1.9
django-rest-swagger==2.2.0
djangorestframework==3.6.3